+++ src/common/jsonconfig.c
@@ -16,6 +16,11 @@ bool ffJsonConfigParseModuleArgs(const char* key, yyjson_val* val, FFModuleArgs*
         ffStrbufSetNS(&moduleArgs->key, (uint32_t) yyjson_get_len(val), yyjson_get_str(val));
         return true;
     }
+    else if(strcasecmp(key, "keyColor") == 0)
+    {
+        ffOptionParseColor(yyjson_get_str(val), &moduleArgs->keyColor);
+        return true;
+    }
     else if(strcasecmp(key, "format") == 0)
     {
         ffStrbufSetNS(&moduleArgs->outputFormat, (uint32_t) yyjson_get_len(val), yyjson_get_str(val));